Sql_LLM_App/requirements.txt
streamlit
google-geneartiveai
python-dotenv


